Create december_12.adb
Learnt something today, in Ada scalar subtypes with static predicates cannot be used as array indices. This precluded the use of look up tables. I produce code of tertiary assignment quality which takes a bit longer; however I got the correct answer for part one on the first. Part two got the correct answer after debugging with the example. The example is not a complete test because it does not test 180 and 270 rotations. Working out rotations from first principles was fun!