Update december_12.adb
I found some more fundamental errors, the definition of North and South was correct for (0, 0) at top left of map but not correct for (0, 0) at the centre of the map. this also made the implementation of 90 and 270 rotations incorrect. The code now represents a usual coordinate system where the positive quadrent (top Right) corresponds with the North East quadrent.