Create december_17.adb
Made a simple error in part one when counting adjacent cells I tested for the Z coordinate in the Y coordinate range causing a run time exception. Adding the W dimension was trivial just copied  Solve to Solve_2 and added the extra iimension. There may be a more elegant generic solution for a proper Ada expert.