Create december_21.adb
Initially I counted the safe ingredients which was not correct. Unfortunately I did not initially retain the original ingredient llists so I had to add that to the Read_List procedure. For part two I just generated allergen list which was in order alphabetical order of ingredient, this is easy but not correct.  Build_Actual resolves allergens to a unique ingredient. The Allergen_List is in the correct order and hence puts the ingredients in the correct order.