Update december_13.adb
Part two solved via the Chinese Remainder theroem. One issue that slowed the solurion was how to deal with the offsets that are larger than the bus id (period). Looking at the example showed me how to deal with the large offset. The Chinese Remainder solution is is in the DJH repository of reusable components.