Update december_16.adb
Finally got part two. The HUGE error was assuming that any ticket that returned a zero error rate in part one was in fact a valid ticked. Some invalid tickets contained 0 one column which is never valid but were otherwise valid. Adding zero to the sum is still zero and does not affect the sum in the error rate!