Day 20 Part 1
Revisiting this problem which I failed to solve whilst in hospital recovering from surgery in 2020.,I changed the approach to be similar to 2021 day 19 Beacon Scanner. I have used 2D matrix transformations to rotate and flip the individual tiles. I have confirmed that there are only 8 orientations, only flip about the x axis is performed as a 90 degree rotation and a y axis flip is the same as an x axis flip. During my development and debug I printed out the length of all the match lists which appeared to be consistent for all tiles, that is, for each tile all orientations matched the same number of times (2, 3 or 4). This was somewhat unexpected; however on further consideration this should be the case because the picture can itself be assembled in eight different orientations,