Day 19 Part 1 only
The solution works by creating the set of all messages satisfying the rules. I came up with this idea in 2020 when the problem was released; however I discarded the idea in the basis that there were likely to be too many, hence the solution would be infeasible.  Subsequently I looked at writing an LR parser, this looked to be difficult given that multiple rules would likely match any state plus a single character or pair of characters. In Megatrends I came across someone using this solution which they said ran in reasonable time. In my case the set of all valid messages is only of the order of 2,000,000 and can be generated in less than 10s!
Some of my implementation choices are not going to be particularly helpful for part two.