Day 19 Completed
Needed help from Megatrends. The observation that rules 31 and 42 returned products that are eight symbols long, plus that valid messages must have at least one instance of a rule 42 product followed by balanced pairs of 42 and 31 are key to the solution. A further observation is the set of 31 products is disjoint from the set of 42 products also the union of both sets contains all possible combinations of two symbols.